Added neovim
Added tests for neovim and app config etc.
Removed console.log from teems, should be handled in CLI wrapper otherwise it's really annoying
Should be backwards compatible as a the nvim app can just throw an error if it can't find the key with the theme name. As every app lives in its own promise chain (would be cool for Goroutines?) nothing else is affected