Fixed bug with transforms
I was filtering the remaining transforms based on equality of the transform itself. Due to not having paid enough attention to the test files, I didn't notice this bug.

The problem is that Xresource for example defines a pretty generic matcher which matches colorX and replaces it with colorX from colors. So I use the same matcher for colors 0 through 15. The first time it found such a matching transform, it would filter all transforms based on whether they were NOT EQUAL to the matched transform. So it would, at once, remove all of the generic transforms because they all look the same.

I am now filtering based on the index instead